+++ sqlstore/src/main/resources/db/migration_beforeMigrate/05_add_containermodel.sql
@@ -4,30 +4,33 @@ DELIMITER //
 
 DROP PROCEDURE IF EXISTS addContainerModel//
 CREATE PROCEDURE addContainerModel(
-  iName varchar(255),
+  iAlias varchar(255),
   iBarcode varchar(255),
   iPartitionCount int(11),
-  iPlatformType varchar(255),
+  iPlatform varchar(255),
   iInstrumentModel varchar(50)
 ) BEGIN
   DECLARE errorMessage varchar(300);
-  IF NOT EXISTS (SELECT 1 FROM Platform WHERE instrumentModel = iPlatformModel) THEN
-    SET errorMessage = CONCAT('Platform ''', iPlatformModel, ''' not found.');
+  IF NOT EXISTS (SELECT 1 FROM InstrumentModel WHERE platform = iPlatform AND alias = iInstrumentModel) THEN
+    SET errorMessage = CONCAT('Instrument model ''', iInstrumentModel, ''' not found.');
     SIGNAL SQLSTATE '45000' SET message_text = errorMessage;
   END IF;
-  IF EXISTS (SELECT 1 FROM SequencingContainerModel WHERE alias = iName AND partitionCount = iPartitionCount 
-  AND platformType = iPlatformType) THEN
-    SET errorMessage = CONCAT('Container model ''', iName, ''' with ', iPartitionCount, ' partitions for platform ',
-    iPlatformType, ' already exists.');
-    SIGNAL SQLSTATE '45000' SET message_text = errorMessage;
+  IF NOT EXISTS (SELECT 1 FROM SequencingContainerModel WHERE alias = iAlias AND partitionCount = iPartitionCount 
+      AND platformType = iPlatform) THEN
+    INSERT INTO SequencingContainerModel (alias, identificationBarcode, partitionCount, platformType)
+    VALUES (iAlias, iBarcode, iPartitionCount, iPlatform);
   END IF;
-  
-  INSERT INTO SequencingContainerModel (alias, identificationBarcode, partitionCount, platformType)
-  VALUES (iName, iBarcode, iPartitionCount, iPlatformType);
-  INSERT INTO SequencingContainerModel_InstrumentModel (sequencingContainerModelId, instrumentModelId)
-  VALUES (
-    (SELECT sequencingContainerModelId FROM SequencingContainerModel WHERE alias = iName),
-    (SELECT instrumentModelId FROM InstrumentModel WHERE alias = iInstrumentModel));
+  IF NOT EXISTS (SELECT 1 FROM SequencingContainerModel_InstrumentModel scmim
+    JOIN SequencingContainerModel scm ON scm.sequencingContainerModelId = scmim.sequencingContainerModelId
+    JOIN InstrumentModel im ON im.instrumentModelId = scmim.instrumentModelId
+    WHERE scm.alias = iAlias AND im.alias = iInstrumentModel
+  ) THEN
+    INSERT INTO SequencingContainerModel_InstrumentModel (sequencingContainerModelId, instrumentModelId)
+    VALUES (
+      (SELECT sequencingContainerModelId FROM SequencingContainerModel WHERE alias = iAlias),
+      (SELECT instrumentModelId FROM InstrumentModel WHERE alias = iInstrumentModel)
+    );
+  END IF;
 END//
 
 DELIMITER ;